pyproject.toml
[build-system]
requires = [
    "setuptools>=77.0.0",
    "wheel",
    "pybind11>=2.10.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "llamatelemetry"
version = "1.0.0"
description = "CUDA-first OpenTelemetry Python SDK for LLM inference observability. GPU-native telemetry for quantized LLM inference pipelines with multi-GPU NCCL, GGUF/llama.cpp runtimes, and real-time graph-based introspection via pygraphistry. Vendor-neutral OTLP export."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Waqas Muhammad", email = "waqasm86@gmail.com"}
]
keywords = [
    "llm", "cuda", "inference", "opentelemetry", "observability", "telemetry",
    "t4", "tesla-t4", "kaggle", "flashattention", "quantization", "gguf",
    "graphistry", "nccl", "deep-learning", "gpu", "llama-cpp", "split-gpu",
    "neural-network-visualization", "tracing", "metrics", "otlp",
    "llm-observability", "gpu-metrics", "inference-tracing"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: C++",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Jupyter",
    "Framework :: Jupyter :: JupyterLab",
    "Environment :: GPU :: NVIDIA CUDA :: 12",
]

# Streamlined core dependencies for v1.0.0
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
gpu = [
    "pynvml>=11.0.0",
]
kaggle = [
    "huggingface_hub>=0.20.0",
    "tqdm>=4.65.0",
]
graphistry = [
    "pygraphistry>=0.32.0",
    "pandas>=1.3.0",
]
extras = [
    "numpy>=1.24.0",
    "ipywidgets>=7.6.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "mypy>=0.950",
    "pybind11>=2.10.0",
]
all = [
    "pynvml>=11.0.0",
    "huggingface_hub>=0.20.0",
    "tqdm>=4.65.0",
    "pygraphistry>=0.32.0",
    "pandas>=1.3.0",
    "numpy>=1.24.0",
    "ipywidgets>=7.6.0",
    "matplotlib>=3.5.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-instrumentation>=0.42.0",
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/llamatelemetry/llamatelemetry"
Documentation = "https://llamatelemetry.github.io/"
Repository = "https://github.com/llamatelemetry/llamatelemetry"
"Bug Tracker" = "https://github.com/llamatelemetry/llamatelemetry/issues"
Changelog = "https://github.com/llamatelemetry/llamatelemetry/blob/main/CHANGELOG.md"

# Package discovery configuration
[tool.setuptools.packages.find]
include = ["llamatelemetry*"]
exclude = [
    "tests*",
    "examples*",
    "docs*",
    "notebooks*",
    "logs*",
    "csrc*",
    "archive*",
    "scripts*"
]

# Exclude package data patterns
[tool.setuptools.exclude-package-data]
llamatelemetry = [
    "binaries/**/*",
    "lib/**/*.so*",
    "models/**/*",
    "*.gguf",
    "*.tar.gz",
    "*.so",
    "*.so.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
